refactor: Workshop 파일 네이밍 및 구조 정리 - 2
- +TopBanner -> TOpSection
- GridHelper -> GridBuilder
- MyItemsView -> WorkshopItemsView (근데 내 아이템뷰 사라지긴함)
- WorkshopPreview -> WorkshopItemDetailView (프리뷰보다 직관적인 네이밍으로 변경)
- 위 파일들에 대한 참조 변경
